Move get_connection_points_for, extract common error message
